+++ Adapter/RabbitMq/AMQPMessage.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Kaliop\QueueingBundle\Adapter\RabbitMq;
+
+use PhpAmqpLib\Message\AMQPMessage as BaseMessage;
+
+class AMQPMessage extends BaseMessage
+{
+    protected $queueName;
+
+    public function setQueueName($queueName)
+    {
+        $this->queueName = $queueName;
+    }
+
+    public function getQueueName()
+    {
+        return $this->queueName;
+    }
+}