+++ Service/MessageConsumer.php
@@ -143,6 +143,7 @@ abstract class MessageConsumer implements ConsumerInterface, MessageConsumerInte
     }
 
     /**
+     * @deprecated should be moved to protected access
      * @return MessageInterface
      */
     public function getCurrentMessage()
@@ -166,12 +167,13 @@ abstract class MessageConsumer implements ConsumerInterface, MessageConsumerInte
         // save the message, in case child class needs it for whacky stuff
         $this->currentMessage = $msg;
 
+        $body = null;
         try {
             $body = $this->decodeMessageBody($msg);
 
             // while at it, emit a message, and allow listeners to prevent further execution
             if ($this->dispatcher) {
-                $event = new MessageReceivedEvent($body, $this);
+                $event = new MessageReceivedEvent($msg, $body, $this);
                 if ($this->dispatcher->dispatch(EventsList::MESSAGE_RECEIVED, $event)->isPropagationStopped()) {
                     return;
                 }
@@ -179,8 +181,11 @@ abstract class MessageConsumer implements ConsumerInterface, MessageConsumerInte
 
             $result = $this->consume($body);
 
+            // q: should we one more try-catch block here, to prevent firing MESSAGE_CONSUMPTION_FAILED if in fact the
+            //    consumption went fine?
+
             if ($this->dispatcher) {
-                $event = new MessageConsumedEvent($body, $result, $this);
+                $event = new MessageConsumedEvent($msg, $body, $this, $result);
                 $this->dispatcher->dispatch(EventsList::MESSAGE_CONSUMED, $event);
             }
 
@@ -192,7 +197,7 @@ abstract class MessageConsumer implements ConsumerInterface, MessageConsumerInte
             }
 
             if ($this->dispatcher) {
-                $event = new MessageConsumptionFailedEvent($body, $e, $this);
+                $event = new MessageConsumptionFailedEvent($msg, $body, $this, $e);
                 $this->dispatcher->dispatch(EventsList::MESSAGE_CONSUMPTION_FAILED, $event);
             }
         }